requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
faiss-cpu
openai
python-dotenv
